fix minor error

+++ .bashrc
@@ -31,7 +31,11 @@ unset file
 # This line is executed upon exiting shell. Currently it tries to kill
 # emacs, be careful though. you may not want this.
 function cleanup {
-  killall emacs;
+  if ! pgrep -x "emacs" > /dev/null
+  then 
+    killall emacs
+  fi
+  sleep 10
 }
 trap cleanup EXIT
